Add WebinarSummary component Move duplicated Code into WebinarSummary Fix Chip-icon console warning
